Fix application freeze on Linux/X11 when viewing "show tracker without clicking" example (#13)
Fix application freeze on Linux/X11 when viewing "show tracker without clicking" example (#13)

This is a fix for issue #13 on github:
https://github.com/oxyplot/oxyplot-gtksharp/issues/13

In the "Show tracker without clicking" graph (in the PlotController
Examples section of the example browser), the application becomes
unresponsive after moving the mouse over the graph area (specifically,
after moving the mouse over the lines on the graph).

This happens because when we move the tracker label (in the
motion-notify-event), we start to receive an endless deluge of
enter-notify-event and leave-notify-event, with the NotifyType of the
event data set to NotifyType.Inferior, indicating that the mouse has
entered or left the plot area via an inferior window (in this case, the
tracker label's eventbox's window).

My solution here is to ignore these events in cases where the notify
type is set to NotifyType.Inferior. The only problem with this would
occur if there are other 'inferior' gdk windows under the plot area. The
only way that I can think of that this might be the case would be for
the user to subclass PlotView. In which case they would be able to
override these event callbacks anyway.